requirements.txt
Django>=4.0.4,<4.1
djangorestframework>=3.13.1